0318/auth/requirements.txt
asgiref==3.7.2
Django==4.2
django-seed==0.3.1
djangorestframework==3.14.0
Faker==24.2.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2024.1
